% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flourish_charts.R
\name{collect_charts}
\alias{collect_charts}
\title{Collect charts from Flourish.}
\usage{
collect_charts(chart_defs, output_dir)
}
\arguments{
\item{chart_defs}{A list of chart definitions, each a list containing:
\itemize{
\item id: The Flourish chart ID
\item filename: The output filename
\item width: The width of the chart
\item height: The height of the chart
\item scale: The scale factor for the chart
}}

\item{output_dir}{The directory to save the charts to}
}
\description{
Collect charts from Flourish.
}
\examples{
\dontrun{
collect_charts(
  chart_defs = list(
    list(id = "123456", filename = "chart1.png", width = 800, height = 600, scale = 2),
    list(id = "789012", filename = "chart2.png", width = 800, height = 600, scale = 2)
  ),
  "output_dir"
)
}
}
